fix(boot.go,gitreceive/*): move the gitreceive logger to inside the gitreceive package
